+++ x-pack/plugins/lists/public/exceptions/components/and_or_badge/rounded_badge.test.tsx
@@ -1,0 +1,25 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React from 'react';
+import { mount } from 'enzyme';
+
+import { RoundedBadge } from './rounded_badge';
+
+describe('RoundedBadge', () => {
+  test('it renders "and" when "type" is "and"', () => {
+    const wrapper = mount(<RoundedBadge type="and" />);
+
+    expect(wrapper.find('[data-test-subj="and-or-badge"]').at(0).text()).toEqual('AND');
+  });
+
+  test('it renders "or" when "type" is "or"', () => {
+    const wrapper = mount(<RoundedBadge type="or" />);
+
+    expect(wrapper.find('[data-test-subj="and-or-badge"]').at(0).text()).toEqual('OR');
+  });
+});
