+++ x-pack/plugins/lists/public/exceptions/components/and_or_badge/translations.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { i18n } from '@kbn/i18n';
+
+export const AND = i18n.translate('xpack.lists.andOrBadge.andLabel', {
+  defaultMessage: 'AND',
+});
+
+export const OR = i18n.translate('xpack.lists.andOrBadge.orLabel', {
+  defaultMessage: 'OR',
+});
