+++ x-pack/plugins/lists/public/exceptions/components/builder/and_badge.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import { EuiFlexItem } from '@elastic/eui';
 import styled from 'styled-components';
 
-import { AndOrBadge } from '../../and_or_badge';
+import { AndOrBadge } from '../and_or_badge';
 
 const MyInvisibleAndBadge = styled(EuiFlexItem)`
   visibility: hidden;
