+++ x-pack/plugins/lists/public/exceptions/components/builder/helpers.test.ts
@@ -5,6 +5,9 @@
  * 2.0.
  */
 
+import { ENTRIES_WITH_IDS } from '../../../../common/constants.mock';
+import { getEntryExistsMock } from '../../../../common/schemas/types/entry_exists.mock';
+import { getExceptionListItemSchemaMock } from '../../../../common/schemas/response/exception_list_item_schema.mock';
 import {
   fields,
   getField,
@@ -37,8 +40,9 @@ import {
 } from '../../../../common';
 import { OperatorOption } from '../autocomplete/types';
 
-import { BuilderEntry, FormattedBuilderEntry } from './types';
+import { BuilderEntry, ExceptionsBuilderExceptionItem, FormattedBuilderEntry } from './types';
 import {
+  getCorrespondingKeywordField,
   getEntryFromOperator,
   getEntryOnFieldChange,
   getEntryOnListChange,
@@ -46,12 +50,21 @@ import {
   getEntryOnMatchChange,
   getEntryOnOperatorChange,
   getFilteredIndexPatterns,
+  getFormattedBuilderEntries,
+  getFormattedBuilderEntry,
   getOperatorOptions,
+  getUpdatedEntriesOnDelete,
+  isEntryNested,
 } from './helpers';
 
 jest.mock('uuid', () => ({
   v4: jest.fn().mockReturnValue('123'),
 }));
+
+const getEntryExistsWithIdMock = (): EntryExists & { id: string } => ({
+  ...getEntryExistsMock(),
+  id: '123',
+});
 
 const getEntryNestedWithIdMock = (): EntryNested & { id: string } => ({
   ...getEntryNestedMock(),
@@ -993,6 +1006,424 @@ describe('Exception builder helpers', () => {
         },
       };
       expect(output).toEqual(expected);
+    });
+  });
+
+  describe('#getFormattedBuilderEntries', () => {
+    test('it returns formatted entry with field undefined if it unable to find a matching index pattern field', () => {
+      const payloadIndexPattern: IIndexPattern = getMockIndexPattern();
+      const payloadItems: BuilderEntry[] = [getEntryMatchWithIdMock()];
+      const output = getFormattedBuilderEntries(payloadIndexPattern, payloadItems);
+      const expected: FormattedBuilderEntry[] = [
+        {
+          correspondingKeywordField: undefined,
+          entryIndex: 0,
+          field: undefined,
+          id: '123',
+          nested: undefined,
+          operator: isOperator,
+          parent: undefined,
+          value: 'some host name',
+        },
+      ];
+      expect(output).toEqual(expected);
+    });
+
+    test('it returns formatted entries when no nested entries exist', () => {
+      const payloadIndexPattern: IIndexPattern = getMockIndexPattern();
+      const payloadItems: BuilderEntry[] = [
+        { ...getEntryMatchWithIdMock(), field: 'ip', value: 'some ip' },
+        { ...getEntryMatchAnyWithIdMock(), field: 'extension', value: ['some extension'] },
+      ];
+      const output = getFormattedBuilderEntries(payloadIndexPattern, payloadItems);
+      const expected: FormattedBuilderEntry[] = [
+        {
+          correspondingKeywordField: undefined,
+          entryIndex: 0,
+          field: {
+            aggregatable: true,
+            count: 0,
+            esTypes: ['ip'],
+            name: 'ip',
+            readFromDocValues: true,
+            scripted: false,
+            searchable: true,
+            type: 'ip',
+          },
+          id: '123',
+          nested: undefined,
+          operator: isOperator,
+          parent: undefined,
+          value: 'some ip',
+        },
+        {
+          correspondingKeywordField: undefined,
+          entryIndex: 1,
+          field: {
+            aggregatable: true,
+            count: 0,
+            esTypes: ['keyword'],
+            name: 'extension',
+            readFromDocValues: true,
+            scripted: false,
+            searchable: true,
+            type: 'string',
+          },
+          id: '123',
+          nested: undefined,
+          operator: isOneOfOperator,
+          parent: undefined,
+          value: ['some extension'],
+        },
+      ];
+      expect(output).toEqual(expected);
+    });
+
+    test('it returns formatted entries when nested entries exist', () => {
+      const payloadIndexPattern: IIndexPattern = getMockIndexPattern();
+      const payloadParent: EntryNested = {
+        ...getEntryNestedWithIdMock(),
+        entries: [{ ...getEntryMatchWithIdMock(), field: 'child' }],
+        field: 'nestedField',
+      };
+      const payloadItems: BuilderEntry[] = [
+        { ...getEntryMatchWithIdMock(), field: 'ip', value: 'some ip' },
+        { ...payloadParent },
+      ];
+
+      const output = getFormattedBuilderEntries(payloadIndexPattern, payloadItems);
+      const expected: FormattedBuilderEntry[] = [
+        {
+          correspondingKeywordField: undefined,
+          entryIndex: 0,
+          field: {
+            aggregatable: true,
+            count: 0,
+            esTypes: ['ip'],
+            name: 'ip',
+            readFromDocValues: true,
+            scripted: false,
+            searchable: true,
+            type: 'ip',
+          },
+          id: '123',
+          nested: undefined,
+          operator: isOperator,
+          parent: undefined,
+          value: 'some ip',
+        },
+        {
+          correspondingKeywordField: undefined,
+          entryIndex: 1,
+          field: {
+            aggregatable: false,
+            esTypes: ['nested'],
+            name: 'nestedField',
+            searchable: false,
+            type: 'string',
+          },
+          id: '123',
+          nested: 'parent',
+          operator: isOperator,
+          parent: undefined,
+          value: undefined,
+        },
+        {
+          correspondingKeywordField: undefined,
+          entryIndex: 0,
+          field: {
+            aggregatable: false,
+            count: 0,
+            esTypes: ['text'],
+            name: 'child',
+            readFromDocValues: false,
+            scripted: false,
+            searchable: true,
+            subType: {
+              nested: {
+                path: 'nestedField',
+              },
+            },
+            type: 'string',
+          },
+          id: '123',
+          nested: 'child',
+          operator: isOperator,
+          parent: {
+            parent: {
+              entries: [
+                {
+                  field: 'child',
+                  id: '123',
+                  operator: OperatorEnum.INCLUDED,
+                  type: OperatorTypeEnum.MATCH,
+                  value: 'some host name',
+                },
+              ],
+              field: 'nestedField',
+              id: '123',
+              type: OperatorTypeEnum.NESTED,
+            },
+            parentIndex: 1,
+          },
+          value: 'some host name',
+        },
+      ];
+      expect(output).toEqual(expected);
+    });
+  });
+
+  describe('#getUpdatedEntriesOnDelete', () => {
+    test('it removes entry corresponding to "entryIndex"', () => {
+      const payloadItem: ExceptionsBuilderExceptionItem = {
+        ...getExceptionListItemSchemaMock(),
+        entries: ENTRIES_WITH_IDS,
+      };
+      const output = getUpdatedEntriesOnDelete(payloadItem, 0, null);
+      const expected: ExceptionsBuilderExceptionItem = {
+        ...getExceptionListItemSchemaMock(),
+        entries: [
+          {
+            field: 'some.not.nested.field',
+            id: '123',
+            operator: OperatorEnum.INCLUDED,
+            type: OperatorTypeEnum.MATCH,
+            value: 'some value',
+          },
+        ],
+      };
+      expect(output).toEqual(expected);
+    });
+
+    test('it removes nested entry of "entryIndex" with corresponding parent index', () => {
+      const payloadItem: ExceptionsBuilderExceptionItem = {
+        ...getExceptionListItemSchemaMock(),
+        entries: [
+          {
+            ...getEntryNestedWithIdMock(),
+            entries: [{ ...getEntryExistsWithIdMock() }, { ...getEntryMatchAnyWithIdMock() }],
+          },
+        ],
+      };
+      const output = getUpdatedEntriesOnDelete(payloadItem, 0, 0);
+      const expected: ExceptionsBuilderExceptionItem = {
+        ...getExceptionListItemSchemaMock(),
+        entries: [
+          { ...getEntryNestedWithIdMock(), entries: [{ ...getEntryMatchAnyWithIdMock() }] },
+        ],
+      };
+      expect(output).toEqual(expected);
+    });
+
+    test('it removes entire nested entry if after deleting specified nested entry, there are no more nested entries left', () => {
+      const payloadItem: ExceptionsBuilderExceptionItem = {
+        ...getExceptionListItemSchemaMock(),
+        entries: [
+          {
+            ...getEntryNestedWithIdMock(),
+            entries: [{ ...getEntryExistsWithIdMock() }],
+          },
+        ],
+      };
+      const output = getUpdatedEntriesOnDelete(payloadItem, 0, 0);
+      const expected: ExceptionsBuilderExceptionItem = {
+        ...getExceptionListItemSchemaMock(),
+        entries: [],
+      };
+      expect(output).toEqual(expected);
+    });
+  });
+
+  describe('#getFormattedBuilderEntry', () => {
+    test('it returns entry with a value for "correspondingKeywordField" when "item.field" is of type "text" and matching keyword field exists', () => {
+      const payloadIndexPattern: IIndexPattern = {
+        ...getMockIndexPattern(),
+        fields: [
+          ...fields,
+          {
+            aggregatable: false,
+            count: 0,
+            esTypes: ['text'],
+            name: 'machine.os.raw.text',
+            readFromDocValues: true,
+            scripted: false,
+            searchable: false,
+            type: 'string',
+          },
+        ],
+      };
+      const payloadItem: BuilderEntry = {
+        ...getEntryMatchWithIdMock(),
+        field: 'machine.os.raw.text',
+        value: 'some os',
+      };
+      const output = getFormattedBuilderEntry(
+        payloadIndexPattern,
+        payloadItem,
+        0,
+        undefined,
+        undefined
+      );
+      const expected: FormattedBuilderEntry = {
+        correspondingKeywordField: getField('machine.os.raw'),
+        entryIndex: 0,
+        field: {
+          aggregatable: false,
+          count: 0,
+          esTypes: ['text'],
+          name: 'machine.os.raw.text',
+          readFromDocValues: true,
+          scripted: false,
+          searchable: false,
+          type: 'string',
+        },
+        id: '123',
+        nested: undefined,
+        operator: isOperator,
+        parent: undefined,
+        value: 'some os',
+      };
+      expect(output).toEqual(expected);
+    });
+
+    test('it returns "FormattedBuilderEntry" with value "nested" of "child" when "parent" and "parentIndex" are defined', () => {
+      const payloadIndexPattern: IIndexPattern = getMockIndexPattern();
+      const payloadItem: BuilderEntry = { ...getEntryMatchWithIdMock(), field: 'child' };
+      const payloadParent: EntryNested = {
+        ...getEntryNestedWithIdMock(),
+        entries: [{ ...getEntryMatchWithIdMock(), field: 'child' }],
+        field: 'nestedField',
+      };
+      const output = getFormattedBuilderEntry(
+        payloadIndexPattern,
+        payloadItem,
+        0,
+        payloadParent,
+        1
+      );
+      const expected: FormattedBuilderEntry = {
+        correspondingKeywordField: undefined,
+        entryIndex: 0,
+        field: {
+          aggregatable: false,
+          count: 0,
+          esTypes: ['text'],
+          name: 'child',
+          readFromDocValues: false,
+          scripted: false,
+          searchable: true,
+          subType: {
+            nested: {
+              path: 'nestedField',
+            },
+          },
+          type: 'string',
+        },
+        id: '123',
+        nested: 'child',
+        operator: isOperator,
+        parent: {
+          parent: {
+            entries: [{ ...payloadItem }],
+            field: 'nestedField',
+            id: '123',
+            type: OperatorTypeEnum.NESTED,
+          },
+          parentIndex: 1,
+        },
+        value: 'some host name',
+      };
+      expect(output).toEqual(expected);
+    });
+
+    test('it returns non nested "FormattedBuilderEntry" when "parent" and "parentIndex" are not defined', () => {
+      const payloadIndexPattern: IIndexPattern = getMockIndexPattern();
+      const payloadItem: BuilderEntry = {
+        ...getEntryMatchWithIdMock(),
+        field: 'ip',
+        value: 'some ip',
+      };
+      const output = getFormattedBuilderEntry(
+        payloadIndexPattern,
+        payloadItem,
+        0,
+        undefined,
+        undefined
+      );
+      const expected: FormattedBuilderEntry = {
+        correspondingKeywordField: undefined,
+        entryIndex: 0,
+        field: {
+          aggregatable: true,
+          count: 0,
+          esTypes: ['ip'],
+          name: 'ip',
+          readFromDocValues: true,
+          scripted: false,
+          searchable: true,
+          type: 'ip',
+        },
+        id: '123',
+        nested: undefined,
+        operator: isOperator,
+        parent: undefined,
+        value: 'some ip',
+      };
+      expect(output).toEqual(expected);
+    });
+  });
+
+  describe('#isEntryNested', () => {
+    test('it returns "false" if payload is not of type EntryNested', () => {
+      const payload: BuilderEntry = getEntryMatchWithIdMock();
+      const output = isEntryNested(payload);
+      const expected = false;
+      expect(output).toEqual(expected);
+    });
+
+    test('it returns "true if payload is of type EntryNested', () => {
+      const payload: EntryNested = getEntryNestedWithIdMock();
+      const output = isEntryNested(payload);
+      const expected = true;
+      expect(output).toEqual(expected);
+    });
+  });
+
+  describe('#getCorrespondingKeywordField', () => {
+    test('it returns matching keyword field if "selectedFieldIsTextType" is true and keyword field exists', () => {
+      const output = getCorrespondingKeywordField({
+        fields,
+        selectedField: 'machine.os.raw.text',
+      });
+
+      expect(output).toEqual(getField('machine.os.raw'));
+    });
+
+    test('it returns undefined if "selectedFieldIsTextType" is false', () => {
+      const output = getCorrespondingKeywordField({
+        fields,
+        selectedField: 'machine.os.raw',
+      });
+
+      expect(output).toEqual(undefined);
+    });
+
+    test('it returns undefined if "selectedField" is empty string', () => {
+      const output = getCorrespondingKeywordField({
+        fields,
+        selectedField: '',
+      });
+
+      expect(output).toEqual(undefined);
+    });
+
+    test('it returns undefined if "selectedField" is undefined', () => {
+      const output = getCorrespondingKeywordField({
+        fields,
+        selectedField: undefined,
+      });
+
+      expect(output).toEqual(undefined);
     });
   });
 });
