+++ x-pack/plugins/lists/public/shared_exports.ts
@@ -39,3 +39,6 @@ export {
   UseExceptionListsSuccess,
 } from './exceptions/types';
 export { BuilderEntryItem } from './exceptions/components/builder/entry_renderer';
+export { BuilderAndBadgeComponent } from './exceptions/components/builder/and_badge';
+export { BuilderEntryDeleteButtonComponent } from './exceptions/components/builder/entry_delete_button';
+export { BuilderExceptionListItemComponent } from './exceptions/components/builder/exception_item_renderer';
