+++ x-pack/plugins/security_solution/public/common/components/exceptions/builder/index.tsx
@@ -9,11 +9,14 @@ import React, { useCallback, useEffect, useReducer } from 'react';
 import { EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import styled from 'styled-components';
 
+import { HttpStart } from 'kibana/public';
+import { AutocompleteStart } from 'src/plugins/data/public';
+import { isEqlRule, isThresholdRule } from '../../../../../common/detection_engine/utils';
 import { addIdToItem } from '../../../../../common';
 import { Type } from '../../../../../common/detection_engine/schemas/common/schemas';
-import { BuilderExceptionListItemComponent } from './exception_item';
 import { IIndexPattern } from '../../../../../../../../src/plugins/data/common';
 import {
+  BuilderExceptionListItemComponent,
   ExceptionListItemSchema,
   NamespaceType,
   exceptionListItemSchema,
@@ -22,7 +25,7 @@ import {
   CreateExceptionListItemSchema,
   ExceptionListType,
   entriesNested,
-} from '../../../../../public/lists_plugin_deps';
+} from '../../../../../public/shared_imports';
 import { AndOrBadge } from '../../and_or_badge';
 import { BuilderLogicButtons } from './logic_buttons';
 import { getNewExceptionItem, filterExceptionItems } from '../helpers';
@@ -66,6 +69,8 @@ interface OnChangeProps {
 }
 
 interface ExceptionBuilderProps {
+  httpService: HttpStart;
+  autocompleteService: AutocompleteStart;
   exceptionListItems: ExceptionsBuilderExceptionItem[];
   listType: ExceptionListType;
   listId: string;
@@ -80,6 +85,8 @@ interface ExceptionBuilderProps {
 }
 
 export const ExceptionBuilderComponent = ({
+  httpService,
+  autocompleteService,
   exceptionListItems,
   listType,
   listId,
@@ -374,6 +381,9 @@ export const ExceptionBuilderComponent = ({
               ))}
             <EuiFlexItem grow={false}>
               <BuilderExceptionListItemComponent
+                allowLargeValueLists={!isEqlRule(ruleType) && !isThresholdRule(ruleType)}
+                httpService={httpService}
+                autocompleteService={autocompleteService}
                 key={getExceptionListItemId(exceptionListItem, index)}
                 exceptionItem={exceptionListItem}
                 indexPattern={indexPatterns}
@@ -385,7 +395,6 @@ export const ExceptionBuilderComponent = ({
                 onChangeExceptionItem={handleExceptionItemChange}
                 onlyShowListOperators={containsValueListEntry(exceptions)}
                 setErrorsExist={setErrorsExist}
-                ruleType={ruleType}
               />
             </EuiFlexItem>
           </EuiFlexGroup>
