+++ x-pack/plugins/security_solution/public/shared_imports.ts
@@ -59,4 +59,7 @@ export {
   addEndpointExceptionList,
   withOptionalSignal,
   BuilderEntryItem,
+  BuilderAndBadgeComponent,
+  BuilderEntryDeleteButtonComponent,
+  BuilderExceptionListItemComponent,
 } from '../../lists/public';
